Some UI work on Send Email page
