--- src/features/emailDapp/EmailDapp.css
@@ -1,24 +1,0 @@
-#search {
-  position: relative;
-  border-radius: 4px;
-  background-color: rgba(255, 255, 255, 0.15);
-  cursor: pointer;
-  transition: background-color 250ms ease-in-out;
-  width: 100%;
-  border: 1px solid #788896;
-}
-
-#iconWrapper {
-  padding: 0 16px;
-  height: 100%;
-  position: absolute;
-  pointer-events: none;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-#inputSearch {
-  padding: 12px 12px 12px 0;
-  padding-left: calc(1em + 32px);
-}
