+++ src/features/index.ts
@@ -3,5 +3,5 @@ export { default as Navigation } from './navigation/Navigation';
 export { default as MyProtectedData } from './myProtectedData/MyProtectedData';
 export { default as NewProtectedData } from './myProtectedData/newProtectedData/NewProtectedData';
 export { default as OneProtectedData } from './myProtectedData/OneProtectedData/OneProtectedData';
-export { default as EmailDapp } from './emailDapp/EmailDapp';
 export { default as SendEmail } from './sendEmail/SendEmail';
+export { default as SendEmailForm } from './sendEmail/SendEmailForm';
