+++ src/features/sendEmail/SendEmail.css
@@ -1,21 +1,24 @@
-#textArea {
-  border-color: #c4c4c4;
+#search {
+  position: relative;
   border-radius: 4px;
-  font-family: 'Roboto', 'Helvetica', 'Arial', sans-serif;
-  font-weight: 400;
-  font-size: 1rem;
-  line-height: 1.4375em;
-  letter-spacing: 0.00938em;
-  color: rgba(0, 0, 0, 0.87);
-  box-sizing: border-box;
+  background-color: rgba(255, 255, 255, 0.15);
+  cursor: pointer;
+  transition: background-color 250ms ease-in-out;
+  width: 100%;
+  border: 1px solid #788896;
 }
 
-#textArea:focus {
-  border-color: blueviolet;
+#iconWrapper {
+  padding: 0 16px;
+  height: 100%;
+  position: absolute;
+  pointer-events: none;
+  display: flex;
+  align-items: center;
+  justify-content: center;
 }
 
-.sendEmailButton {
-  width: 80px;
-  margin: auto;
-  margin-right: 0;
+#inputSearch {
+  padding: 12px 12px 12px 0;
+  padding-left: calc(1em + 32px);
 }
