+++ src/features/sendEmail/SendEmail.tsx
@@ -1,172 +1,140 @@
-import { useEffect, useState } from 'react';
-import {
-  Alert,
-  Box,
-  Snackbar,
-  TextField,
-  TextareaAutosize,
-} from '@mui/material';
-import FormControl from '@mui/material/FormControl';
-import Select, { SelectChangeEvent } from '@mui/material/Select';
-import MenuItem from '@mui/material/MenuItem';
-import InputLabel from '@mui/material/InputLabel';
+import { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { Address, Contact, TimeStamp } from '@iexec/web3mail';
+import SearchIcon from '@mui/icons-material/Search';
+import { Box, InputBase } from '@mui/material';
+import { DataGrid, GridColDef } from '@mui/x-data-grid';
+import SendIcon from '@mui/icons-material/Send';
+import { useAccount } from 'wagmi';
 import { Button } from '@/components/ui/button';
+import { selectAppIsConnected, useFetchMyContactsQuery } from '@/app/appSlice';
+import { useAppSelector } from '@/app/hooks';
+import { getLocalDateFromTimeStamp } from '@/utils/utils';
 import './SendEmail.css';
-import { useParams } from 'react-router-dom';
-import { useSendEmailMutation } from '@/app/appSlice';
 
-const MAX_CHARACTERS_SENDER_NAME = 20;
-const MAX_CHARACTERS_MESSAGE_SUBJECT = 78;
+type Row = {
+  id: string;
+  owner: Address;
+  protectedDataAddress: Address;
+  accessGrantTimestamp: TimeStamp;
+};
 
 export default function SendEmail() {
-  const { receiverAddress, protectedDataAddress } = useParams();
+  const navigate = useNavigate();
+  const { address } = useAccount();
+  const isAccountConnected = useAppSelector(selectAppIsConnected);
 
-  //RTK Mutation hook
-  const [sendEmail, result] = useSendEmailMutation();
-
-  //for textarea
-  const [message, setMessage] = useState('');
-  const charactersRemainingMessage = 512000 - message.length;
-
-  //for name et dataType
-  const [messageSubject, setMessageSubject] = useState('');
-  const charactersRemainingSubject =
-    MAX_CHARACTERS_MESSAGE_SUBJECT - messageSubject.length;
-
-  const [contentType, setContentType] = useState('text/plain');
-
-  const [senderName, setSenderName] = useState('');
-  const charactersRemainingSenderName =
-    MAX_CHARACTERS_SENDER_NAME - senderName.length;
-
-  //handle functions
-  const handleMessageSubjectChange = (event: any) => {
-    const inputValue = event.target.value;
-    setMessageSubject(inputValue);
-  };
-
-  const handleChange = (event: any) => {
-    const inputValue = event.target.value;
-    setMessage(inputValue);
-  };
-
-  const sendEmailHandle = () => {
-    if (!protectedDataAddress) return;
-    sendEmail({
-      senderName,
-      contentType,
-      emailSubject: messageSubject,
-      emailContent: message,
-      protectedData: protectedDataAddress,
-    });
-  };
-
-  //snackbar notification
-  const [open, setOpen] = useState(false);
-  const [success, setSuccess] = useState(false);
-
-  useEffect(() => {
-    if (result.isSuccess || result.isError) {
-      setSuccess(result.isSuccess);
-      setOpen(true);
+  //query RTK API as query hook
+  const { data: myContacts = [], isLoading } = useFetchMyContactsQuery(
+    address as string,
+    {
+      skip: !isAccountConnected,
     }
-  }, [result]);
+  );
 
-  const handleClose = (_: React.SyntheticEvent | Event, reason?: string) => {
-    if (reason === 'clickaway') {
-      return;
-    }
-    setOpen(false);
-  };
+  //for search bar
+  const [searchTerm, setSearchTerm] = useState('');
 
-  const handleSelectContentType = (event: SelectChangeEvent) => {
-    setContentType(event.target.value as string);
-  };
-
-  const handleSenderNameChange = (event: any) => {
-    const inputValue = event.target.value;
-    setSenderName(inputValue);
-  };
-  return (
-    <Box sx={{ m: 6, mx: 20 }}>
-      <h2>Send Mail to {receiverAddress}</h2>
-      <Box sx={{ my: 2, display: 'flex', flexDirection: 'column' }}>
-        <TextField
-          fullWidth
-          id="sender-name"
-          label="Sender name"
-          variant="outlined"
-          value={senderName}
-          onChange={handleSenderNameChange}
-          className="mt-6"
-          inputProps={{ maxLength: MAX_CHARACTERS_SENDER_NAME }}
-        />
-        <p className="italic text-sm my-2">
-          {charactersRemainingSenderName >= 0
-            ? charactersRemainingSenderName
-            : 0}{' '}
-          {charactersRemainingSenderName > 1 ? 'characters' : 'character'}{' '}
-          remaining
-        </p>
-        <TextField
-          fullWidth
-          id="Message subject"
-          label="Message subject"
-          variant="outlined"
-          required
-          value={messageSubject}
-          onChange={handleMessageSubjectChange}
-          inputProps={{ maxLength: MAX_CHARACTERS_MESSAGE_SUBJECT }}
-          className="mt-6"
-        />
-        <p className="italic text-sm my-2">
-          {charactersRemainingSubject >= 0 ? charactersRemainingSubject : 0}{' '}
-          {charactersRemainingSubject > 1 ? 'characters' : 'character'}{' '}
-          remaining
-        </p>
-        <FormControl sx={{ textAlign: 'left', mt: 3 }} fullWidth>
-          <InputLabel id="content-type-label">Content Type</InputLabel>
-          <Select
-            labelId="content-type-label"
-            id="content-type-select"
-            value={contentType}
-            label="Content type"
-            onChange={handleSelectContentType}
-          >
-            <MenuItem value="text/plain">text/plain</MenuItem>
-            <MenuItem value="text/html">text/html</MenuItem>
-          </Select>
-        </FormControl>
-        <TextareaAutosize
-          required
-          placeholder="Enter email content"
-          value={message}
-          onChange={handleChange}
-          id="textArea"
-          className="w-full border mt-4 !h-[200px] p-3"
-        />
-        <p className="italic text-sm my-2">
-          {charactersRemainingMessage} characters remaining
-        </p>
-        <Button className="sendEmailButton" onClick={sendEmailHandle}>
-          Send
-        </Button>
-      </Box>
-
-      <Snackbar
-        open={open}
-        autoHideDuration={6000}
-        onClose={handleClose}
-        anchorOrigin={{ horizontal: 'right', vertical: 'top' }}
-      >
-        <Alert
-          onClose={handleClose}
-          severity={success ? 'success' : 'error'}
-          sx={{ width: '100%' }}
+  const columns: GridColDef[] = [
+    {
+      field: 'owner',
+      headerName: 'Eth Address',
+      type: 'string',
+      flex: 2,
+    },
+    {
+      field: 'protectedDataAddress',
+      headerName: 'Protected Data',
+      type: 'string',
+      flex: 2,
+    },
+    { field: 'accessGrantTimestamp', headerName: 'Subscribed on', flex: 1 },
+    {
+      field: 'Actions',
+      headerName: 'Actions',
+      sortable: false,
+      width: 175,
+      renderCell: (params) => (
+        <Button
+          size="sm"
+          className="pl-3.5"
+          onClick={() =>
+            navigate(`./${params.row.owner}/${params.row.protectedDataAddress}`)
+          }
         >
-          {success ? 'The email has been sent!' : 'Failed to send the Email!'}
-        </Alert>
-      </Snackbar>
+          <SendIcon className="text-sm" />
+          <span className="pl-1.5">Send web3 email</span>
+        </Button>
+      ),
+    },
+  ];
+
+  //modified the return of fetchMyContact in the web3Mail SDK to order them by timestamp
+  const rows: Row[] = [...myContacts]
+    .sort(
+      (a: Contact, b: Contact) =>
+        Date.parse(b.accessGrantTimestamp) - Date.parse(a.accessGrantTimestamp)
+    )
+    .map((contact: Contact, index: number) => {
+      return {
+        id: index.toString(),
+        owner: contact.owner.toLowerCase(),
+        protectedDataAddress: contact.address.toLowerCase(),
+        accessGrantTimestamp: getLocalDateFromTimeStamp(
+          contact.accessGrantTimestamp
+        ),
+      };
+    });
+
+  const filteredRows: Row[] = rows.filter((row: { owner: string }) =>
+    row.owner.toLowerCase().includes(searchTerm.toLowerCase())
+  );
+
+  const handleSearchChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setSearchTerm(e.target.value);
+  };
+
+  return (
+    <Box>
+      <h2>Contact List</h2>
+      <p>
+        Contacts who have protected their email address data and have allowed
+        you to use it. You can send them an email, without having access to
+        their email address.
+      </p>
+      <Box id="search" sx={{ mt: 5 }}>
+        <div id="iconWrapper">
+          <SearchIcon sx={{ color: '#788896' }} />
+        </div>
+        <InputBase
+          placeholder="Search address"
+          inputProps={{ 'aria-label': 'search' }}
+          id="inputSearch"
+          sx={{ width: '100%' }}
+          value={searchTerm}
+          onChange={handleSearchChange}
+        />
+      </Box>
+      {filteredRows.length > 0 && !isLoading && (
+        <div className="mt-8 w-full">
+          <DataGrid
+            disableColumnMenu
+            rows={filteredRows}
+            columns={columns}
+            sx={{ border: 'none' }}
+          />
+        </div>
+      )}
+      {filteredRows.length === 0 && !isLoading && (
+        <Box sx={{ textAlign: 'center', my: 5 }}>
+          <h4>You have no subscribers!</h4>
+        </Box>
+      )}
+      {isLoading && (
+        <Box sx={{ textAlign: 'center', my: 5 }}>
+          <h4>Fetching your contacts...</h4>
+        </Box>
+      )}
     </Box>
   );
 }
