+++ src/features/sendEmail/SendEmailForm.css
@@ -1,0 +1,21 @@
+#textArea {
+  border-color: #c4c4c4;
+  border-radius: 4px;
+  font-family: 'Roboto', 'Helvetica', 'Arial', sans-serif;
+  font-weight: 400;
+  font-size: 1rem;
+  line-height: 1.4375em;
+  letter-spacing: 0.00938em;
+  color: rgba(0, 0, 0, 0.87);
+  box-sizing: border-box;
+}
+
+#textArea:focus {
+  border-color: blueviolet;
+}
+
+.sendEmailButton {
+  width: 80px;
+  margin: auto;
+  margin-right: 0;
+}
